changed formatting, modified get_all_bookmarks
