+++ .gitignore
@@ -1,1 +1,2 @@
 project3
+__pycache__
